[query] Encapsulate TabixLineIterator buffering logic
Split it out into BGZipLineReader.
