+++ hail/src/main/scala/is/hail/io/tabix/TabixReader.scala
@@ -1,11 +1,10 @@
 package is.hail.io.tabix
 
 import java.io.InputStream
-import java.nio.charset.StandardCharsets
 
 import htsjdk.samtools.util.FileExtensions
 import htsjdk.tribble.util.ParsingUtils
-import is.hail.io.compress.BGzipInputStream
+import is.hail.io.compress.BGZipLineReader
 import is.hail.io.fs.FS
 import is.hail.utils._
 import is.hail.backend.BroadcastValue
@@ -70,23 +69,6 @@ object TabixReader {
     ((is.read() & 0xff).asInstanceOf[Long] << 40) |
     ((is.read() & 0xff).asInstanceOf[Long] << 48) |
     ((is.read() & 0xff).asInstanceOf[Long] << 56)
-
-  def readLine(is: InputStream): String = readLine(is, DefaultBufferSize)
-
-  def readLine(is: InputStream, initialBufferSize: Int): String = {
-    val buf = new ArrayBuilder[Byte](initialBufferSize)
-    var c = is.read()
-    while (c >= 0 && c != '\n') {
-      if (c == '\r') {
-        c = is.read()
-        if (c != '\n') buf += '\r'
-      } else {
-        buf += c.asInstanceOf[Byte]
-        c = is.read()
-      }
-    }
-    new String(buf.result(), StandardCharsets.UTF_8)
-  }
 }
 
 class TabixReader(val filePath: String, fs: FS, idxFilePath: Option[String] = None) {
@@ -319,130 +301,23 @@ final class TabixLineIterator(
 {
   private var i: Int = -1
   private var isEof = false
-  private var is = new BGzipInputStream(fsBc.value.openNoCompression(filePath))
-
-  // The line iterator buffer and associated state, we use this to avoid making
-  // endless calls to read() (the no arg version returning int) on the input
-  // stream, we start with 64k to make to make sure that we can always read
-  // a block, and then grow it as neccessary if a line is too large.
-  //
-  // The key invariant here is that bufferCursor should, except possibly in
-  // readLine itself, be less than bufferLen. Should bufferCursor be greater
-  // than or equal to bufferLen, we call refreshBuffer, which also saves the
-  // current input stream virtual offset. If we uphold this invariant, then the
-  // bufferCursor to bufferLen range within the buffer will always contain data
-  // from the current block we are reading, meaning getVirtualOffset will
-  // always return the same value as if we were calling is.read() to read the
-  // lines and is.getVirtualOffset to update the current file pointer for every
-  // call to next
-  private var buffer = new Array[Byte](1 << 16)
-  private var bufferCursor: Int = 0
-  private var bufferLen: Int = 0
-  private var bufferEOF: Boolean = false
-  private var bufferPositionAtLastRead = 0L
-
-  private var virtualFileOffsetAtLastRead = 0L
-
-  private def getVirtualOffset: Long = {
-    virtualFileOffsetAtLastRead + (bufferCursor - bufferPositionAtLastRead)
-  }
-
-  private def virtualSeek(l: Long): Unit = {
-    is.virtualSeek(l)
-    refreshBuffer(0)
-    bufferCursor = 0
-  }
-
-  private def refreshBuffer(start: Int): Unit = {
-    assert(start < buffer.length)
-    bufferLen = start
-    virtualFileOffsetAtLastRead = is.getVirtualOffset
-    bufferPositionAtLastRead = start
-    val bytesRead = is.read(buffer, start, buffer.length - start)
-    if (bytesRead < 0)
-      bufferEOF = true
-    else
-      bufferLen = start + bytesRead
-    bufferCursor = start
-  }
-
-  def decodeString(start: Int, end: Int): String = {
-    var stop = end
-    if (stop > start && buffer(stop) == '\r')
-      stop -= 1
-    val len = end - start
-    new String(buffer, start, len, StandardCharsets.UTF_8)
-  }
-
-  def readLine(): String = {
-    assert(bufferCursor <= bufferLen)
-    if (isEof)
-      return null
-
-    var start = bufferCursor
-
-    while (true) {
-      var str: String = null;
-      while (bufferCursor < bufferLen && buffer(bufferCursor) != '\n') {
-        bufferCursor += 1
-      }
-
-      if (bufferCursor == bufferLen) { // no newline before end of buffer
-        if (bufferEOF) {
-          // `is` indicates end of file
-          isEof = true
-          str = decodeString(start, bufferCursor)
-        } else if (bufferLen == buffer.length) {
-          // line overflows buffer, need to increase buffer size
-          val tmp = new Array[Byte](buffer.length * 2)
-          System.arraycopy(buffer, 0, tmp, 0, buffer.length)
-          buffer = tmp
-          refreshBuffer(bufferLen)
-        } else if (start > 0) {
-          // line does not overflow buffer, but spans buffer.
-          // Copy line left to the beginning of buffer and continue
-          val nToCopy = bufferLen - start
-          System.arraycopy(buffer, start, buffer, 0, nToCopy)
-          start = 0
-          refreshBuffer(nToCopy)
-        } else {
-          // line does not span buffer, but does not fill buffer either, read more
-          refreshBuffer(bufferCursor)
-        }
-      } else { // found a newline
-        str = decodeString(start, bufferCursor)
-      }
-
-      if (str != null) {
-        bufferCursor += 1
-        // we refresh here to make sure that the cursor is never pointing to
-        // one past the end of a block, making it so getVirtualOffset will
-        // never return the pointer pointing past the end of a block (the one
-        // that overlaps with the start of the next block).
-        if (bufferCursor >= bufferLen) {
-          refreshBuffer(0)
-        }
-        return str
-      }
-    }
-    throw new AssertionError()
-  }
+  private var lines = new BGZipLineReader(fsBc, filePath)
 
   def next(): String = {
     var s: String = null
     while (s == null && !isEof) {
-      val curOff = getVirtualOffset
+      val curOff = lines.getVirtualOffset
       if (i < 0 || curOff == 0 || !TbiOrd.less64(curOff, offsets(i)._2)) { // jump to next chunk
         if (i == offsets.length - 1) {
           isEof = true
           return s
         }
         if (i < 0 || offsets(i)._2 != offsets(i + 1)._1) {
-          virtualSeek(offsets(i + 1)._1)
+          lines.virtualSeek(offsets(i + 1)._1)
         }
         i += 1
       }
-      s = readLine()
+      s = lines.readLine()
       if (s != null) {
         if (s.isEmpty || s.charAt(0) == '#')
           s = null // continue
@@ -453,9 +328,9 @@ final class TabixLineIterator(
   }
 
   override def close() {
-    if (is != null) {
-      is.close()
-      is = null
+    if (lines != null) {
+      lines.close()
+      lines = null
     }
   }
 }
